# index/pilosa: batch boltdb and pilosa writes