--- a/sql/index/pilosa/mapping.go
+++ b/sql/index/pilosa/mapping.go
@@ -26,6 +26,11 @@ type mapping struct {
 	mut sync.RWMutex
 	db  *bolt.DB
 
+	// in create mode there's only one transaction closed explicitly by
+	// commit function
+	create bool
+	tx     *bolt.Tx
+
 	clientMut sync.Mutex
 	clients   int
 }
@@ -35,9 +40,15 @@ func newMapping(dir string) *mapping {
 }
 
 func (m *mapping) open() {
+	m.openCreate(false)
+}
+
+// openCreate opens and sets creation mode in the database.
+func (m *mapping) openCreate(create bool) {
 	m.clientMut.Lock()
 	defer m.clientMut.Unlock()
 	m.clients++
+	m.create = create
 }
 
 func (m *mapping) close() error {
@@ -93,6 +104,75 @@ func (m *mapping) rowID(frameName string, value interface{}) (uint64, error) {
 	return binary.LittleEndian.Uint64(val), err
 }
 
+// commit saves current transaction, if cont is true a new transaction will be
+// created again in the next query. Only for create mode.
+func (m *mapping) commit(cont bool) error {
+	m.clientMut.Lock()
+	defer m.clientMut.Unlock()
+
+	var err error
+	if m.create && m.tx != nil {
+		err = m.tx.Commit()
+	}
+
+	m.create = cont
+	m.tx = nil
+
+	return err
+}
+
+func (m *mapping) rollback() error {
+	m.clientMut.Lock()
+	defer m.clientMut.Unlock()
+
+	var err error
+	if m.create && m.tx != nil {
+		err = m.tx.Rollback()
+	}
+
+	m.create = false
+	m.tx = nil
+
+	return err
+}
+
+func (m *mapping) transaction(writable bool, f func(*bolt.Tx) error) error {
+	var tx *bolt.Tx
+	var err error
+	if m.create {
+		m.clientMut.Lock()
+		if m.tx == nil {
+			m.tx, err = m.db.Begin(true)
+			if err != nil {
+				m.clientMut.Unlock()
+				return err
+			}
+		}
+
+		m.clientMut.Unlock()
+
+		tx = m.tx
+	} else {
+		tx, err = m.db.Begin(writable)
+		if err != nil {
+			return err
+		}
+	}
+
+	err = f(tx)
+
+	if m.create {
+		return err
+	}
+
+	if err != nil {
+		tx.Rollback()
+		return err
+	}
+
+	return tx.Commit()
+}
+
 func (m *mapping) getRowID(frameName string, value interface{}) (uint64, error) {
 	var id uint64
 	err := m.query(func() error {
@@ -103,7 +183,7 @@ func (m *mapping) getRowID(frameName string, value interface{}) (uint64, error) 
 			return err
 		}
 
-		err = m.db.Update(func(tx *bolt.Tx) error {
+		err = m.transaction(true, func(tx *bolt.Tx) error {
 			b, err := tx.CreateBucketIfNotExists([]byte(frameName))
 			if err != nil {
 				return err
@@ -116,7 +196,7 @@ func (m *mapping) getRowID(frameName string, value interface{}) (uint64, error) 
 				return nil
 			}
 
-			id = uint64(b.Stats().KeyN)
+			id, _ = b.NextSequence()
 			val = make([]byte, 8)
 			binary.LittleEndian.PutUint64(val, id)
 			err = b.Put(key, val)
@@ -131,7 +211,7 @@ func (m *mapping) getRowID(frameName string, value interface{}) (uint64, error) 
 
 func (m *mapping) putLocation(indexName string, colID uint64, location []byte) error {
 	return m.query(func() error {
-		return m.db.Update(func(tx *bolt.Tx) error {
+		return m.transaction(true, func(tx *bolt.Tx) error {
 			b, err := tx.CreateBucketIfNotExists([]byte(indexName))
 			if err != nil {
 				return err
@@ -192,7 +272,7 @@ func (m *mapping) getLocation(indexName string, colID uint64) ([]byte, error) {
 	var location []byte
 
 	err := m.query(func() error {
-		err := m.db.View(func(tx *bolt.Tx) error {
+		err := m.transaction(true, func(tx *bolt.Tx) error {
 			b := tx.Bucket([]byte(indexName))
 			if b == nil {
 				return fmt.Errorf("bucket %s not found", indexName)
@@ -214,7 +294,7 @@ func (m *mapping) getLocation(indexName string, colID uint64) ([]byte, error) {
 func (m *mapping) getLocationN(indexName string) (int, error) {
 	var n int
 	err := m.query(func() error {
-		err := m.db.View(func(tx *bolt.Tx) error {
+		err := m.transaction(false, func(tx *bolt.Tx) error {
 			b := tx.Bucket([]byte(indexName))
 			if b == nil {
 				return fmt.Errorf("Bucket %s not found", indexName)
@@ -240,7 +320,7 @@ func (m *mapping) get(name string, key interface{}) ([]byte, error) {
 			return err
 		}
 
-		err = m.db.View(func(tx *bolt.Tx) error {
+		err = m.transaction(true, func(tx *bolt.Tx) error {
 			b := tx.Bucket([]byte(name))
 			if b != nil {
 				value = b.Get(buf.Bytes())