Writes are now only done once every 10000 rows. This improves index creation times.

pilosa uses import methods instead of queries to set bits. This occupies less memory and is faster than making batched queries.

boltdb mapping now has an extra mode of working (create mode). When this is active there's only one transaction per batch instead of one transaction per operation.

```
create index commit_hash on commits using pilosa (commit_hash);
```

| num repos | before | now |
|--------------|--|-----|
| 9 | 42 s | 5 s |
| 100 | 8 minutes | 21 s |